[AIRFLOW-1889] Split sensors into separate files
Moving the sensors to seperate files increases
readability of the
code. Also this reduces the code in the big
core.py file.

Closes #2875 from Fokko/AIRFLOW-1889-move-sensors-
to-separate-package
