+++ .gitignore
@@ -112,6 +112,7 @@ ENV/
 
 # PyCharm
 .idea/
+*.iml
 
 # Visual Studio Code
 .vscode/
