+++ airflow/contrib/sensors/bash_sensor.py
@@ -17,7 +17,7 @@ import os
 from subprocess import Popen, STDOUT, PIPE
 from tempfile import gettempdir, NamedTemporaryFile
 from airflow.utils.decorators import apply_defaults
-from airflow.operators.sensors import BaseSensorOperator
+from airflow.sensors.base_sensor_operator import BaseSensorOperator
 from airflow.utils.file import TemporaryDirectory
