+++ airflow/contrib/sensors/emr_step_sensor.py
@@ -33,11 +33,11 @@ class EmrStepSensor(EmrBaseSensor):
     template_ext = ()
 
     @apply_defaults
-    def __init__(
-            self,
-            job_flow_id,
-            step_id,
-            *args, **kwargs):
+    def __init__(self,
+                 job_flow_id,
+                 step_id,
+                 *args,
+                 **kwargs):
         super(EmrStepSensor, self).__init__(*args, **kwargs)
         self.job_flow_id = job_flow_id
         self.step_id = step_id
