+++ airflow/contrib/sensors/pubsub_sensor.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from airflow.contrib.hooks.gcp_pubsub_hook import PubSubHook
-from airflow.operators.sensors import BaseSensorOperator
+from airflow.sensors.base_sensor_operator import BaseSensorOperator
 from airflow.utils.decorators import apply_defaults
