+++ airflow/operators/__init__.py
@@ -62,21 +62,6 @@ _operators = {
     'sqlite_operator': ['SqliteOperator'],
     'mysql_to_hive': ['MySqlToHiveTransfer'],
     'postgres_operator': ['PostgresOperator'],
-    'sensors': [
-        'BaseSensorOperator',
-        'ExternalTaskSensor',
-        'HdfsSensor',
-        'HivePartitionSensor',
-        'HttpSensor',
-        'MetastorePartitionSensor',
-        'NamedHivePartitionSensor',
-        'S3KeySensor',
-        'S3PrefixSensor',
-        'SqlSensor',
-        'TimeDeltaSensor',
-        'TimeSensor',
-        'WebHdfsSensor',
-    ],
     'subdag_operator': ['SubDagOperator'],
     'hive_stats_operator': ['HiveStatsCollectionOperator'],
     's3_to_hive_operator': ['S3ToHiveTransfer'],
