+++ airflow/sensors/time_sensor.py
@@ -1,0 +1,35 @@
+# -*- coding: utf-8 -*-
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from airflow.sensors.base_sensor_operator import BaseSensorOperator
+from airflow.utils import timezone
+from airflow.utils.decorators import apply_defaults
+
+
+class TimeSensor(BaseSensorOperator):
+    """
+    Waits until the specified time of the day.
+
+    :param target_time: time after which the job succeeds
+    :type target_time: datetime.time
+    """
+
+    @apply_defaults
+    def __init__(self, target_time, *args, **kwargs):
+        super(TimeSensor, self).__init__(*args, **kwargs)
+        self.target_time = target_time
+
+    def poke(self, context):
+        self.log.info('Checking if the time (%s) has come', self.target_time)
+        return timezone.utcnow().time() > self.target_time
