+++ tests/contrib/sensors/test_ftp_sensor.py
@@ -13,8 +13,8 @@
 # limitations under the License.
 
 import unittest
-from ftplib import error_perm
 
+from ftplib import error_perm
 from mock import MagicMock
 
 from airflow.contrib.hooks.ftp_hook import FTPHook
