+++ tests/contrib/sensors/test_wasb_sensor.py
@@ -13,8 +13,9 @@
 # limitations under the License.
 #
 
-import datetime
 import unittest
+
+import datetime
 
 from airflow import DAG, configuration
 from airflow.contrib.sensors.wasb_sensor import WasbBlobSensor
@@ -30,7 +31,6 @@ except ImportError:
 
 
 class TestWasbBlobSensor(unittest.TestCase):
-
     _config = {
         'container_name': 'container',
         'blob_name': 'blob',
@@ -83,7 +83,6 @@ class TestWasbBlobSensor(unittest.TestCase):
 
 
 class TestWasbPrefixSensor(unittest.TestCase):
-
     _config = {
         'container_name': 'container',
         'prefix': 'prefix',
@@ -133,6 +132,7 @@ class TestWasbPrefixSensor(unittest.TestCase):
         mock_instance.check_for_prefix.assert_called_once_with(
             'container', 'prefix', timeout=2
         )
+
 
 if __name__ == '__main__':
     unittest.main()
